+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 
 - Editor
   - Add support to rename namespace of namespaced keywords like re-frame events/subs. #978
+  - Improve performance of find-declaration feature. #1021
 
 ## 2022.05.31-17.35.50